add ownership to courses and reviews

+++ src/controllers/course.controller.js
@@ -1,5 +1,6 @@
 const asyncHandler = require('../middlewares/async.middleware')
 const coureService = require('../services/courses.service')
+const bootcampService = require('../services/bootcamps.service')
 
 /**
  * @desc  Get all courses associated with a bootcamp
@@ -21,7 +22,7 @@ exports.getCourses = asyncHandler(async (req, res, next) => {
  * @access Public
  */
 exports.getCourse = asyncHandler(async (req, res, next) => {
-  const course = await coureService.findOne(req.params.id)
+  const course = await coureService.find(req.params.id)
 
   res.status(200).json({
     course,
@@ -35,6 +36,18 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
  */
 exports.createCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId
+  req.body.user = req.user.id
+
+  const bootcamp = await bootcampService.findById(req.params.bootcampId)
+
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update this bootcamp`,
+        401
+      )
+    )
+  }
 
   const course = await coureService.create(req.body)
 
@@ -49,7 +62,18 @@ exports.createCourse = asyncHandler(async (req, res, next) => {
  * @access Private
  */
 exports.updateCourse = asyncHandler(async (req, res, next) => {
-  const course = await coureService.update(req.params.id, req.body)
+  let course = await coureService.findOne(req.params.id)
+
+  if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update this bootcamp`,
+        401
+      )
+    )
+  }
+
+  course = await coureService.update(req.params.id, req.body)
 
   res.status(200).json({
     course,
@@ -62,7 +86,18 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
  * @access Private
  */
 exports.deleteCourse = asyncHandler(async (req, res, next) => {
-  const course = await coureService.delete(req.params.id)
+  let course = await coureService.findOne(req.params.id)
+
+  if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update this bootcamp`,
+        401
+      )
+    )
+  }
+
+  course = await coureService.delete(req.params.id)
 
   res.status(200).json({
     course,
