+++ src/controllers/review.controller.js
@@ -21,6 +21,8 @@ exports.getReviews = asyncHandler(async (req, res, next) => {
  */
 exports.createReview = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId
+  req.body.user = req.user.id
+
   const review = await reviewService.create(req.body)
 
   res.status(201).json({
@@ -34,7 +36,18 @@ exports.createReview = asyncHandler(async (req, res, next) => {
  * @access  Private
  */
 exports.updateReview = asyncHandler(async (req, res, next) => {
-  const review = await reviewService.update(req.params.id, req.body)
+  let review = await reviewService.findById(req.params.id)
+
+  if (review.user.toString() !== req.user.id) {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update this bootcamp`,
+        401
+      )
+    )
+  }
+
+  review = await reviewService.update(req.params.id, req.body)
 
   res.status(200).json({
     review,
@@ -47,7 +60,18 @@ exports.updateReview = asyncHandler(async (req, res, next) => {
  * @access  Private
  */
 exports.deleteReview = asyncHandler(async (req, res, next) => {
-  const review = await reviewService.delete(req.params.id)
+  let review = await reviewService.findById(req.params.id)
+
+  if (review.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update this bootcamp`,
+        401
+      )
+    )
+  }
+
+  review = await reviewService.delete(req.params.id)
 
   res.status(200).json({
     review,
